dp_core: Introduce offloads module
Smart nics may have ability to parse and act the Vrouter tunneled packets.

So, according to the Vrouter tunnels, aka MPLSoGRE, MPLSoUDP, and VXLAN, the
inner packet 5-tupples and the outer packet specifications, the HW can do
some actions. For example, the HW can detect MPLSoGRE packets and to do RSS
on the inner packet 5-tupples or just tag the packets based on the packet HW parser.

Using this HW feature the Vrouter parformance can be improved.

Add a Vrouter offloads module to use the described HW ability.

Signed-off-by: matan <matan@mellanox.com>