btable: optimize entry address calculation
The btable uses division and modulo calculations, which are quiet
expensive in the most platforms, for the entry address calculations.

In case the allocation limit of a btable is power of 2 number,
the calculations can be optimized to use simple bit calculations.

Optimize the entry address calculations in the aforementioned case.

Signed-off-by: Matan Azrad <matan@mellanox.com>